cleanup(handle_manager): unused file removed
